refactor(memory): split ORM models by resource
