+++ backend/database/models/memory/__init__.py
@@ -1,15 +1,14 @@
-from backend.database.models.memory.items import (
+from backend.database.models.memory.context_notes import (
     ContextNote,
     ContextNoteVersion,
-    EventVersion,
-    FactVersion,
-    MemoryItem,
-    MemorySearchDocument,
-    MemoryVersion,
-    StorylineVersion,
-    TriggerVersion,
 )
+from backend.database.models.memory.events import EventVersion
+from backend.database.models.memory.facts import FactVersion
+from backend.database.models.memory.items import MemoryItem, MemoryVersion
 from backend.database.models.memory.revisions import CurrentRevision, MemoryRevision
+from backend.database.models.memory.search_documents import MemorySearchDocument
+from backend.database.models.memory.storylines import StorylineVersion
+from backend.database.models.memory.triggers import TriggerVersion
 
 __all__ = [
     "ContextNote",
