+++ backend/database/models/memory/storylines.py
@@ -1,0 +1,48 @@
+from __future__ import annotations
+
+from uuid import UUID
+
+from sqlalchemy import ForeignKey, SmallInteger, Text, text
+from sqlalchemy.dialects.postgresql import ARRAY, JSONB, UUID as PGUUID
+from sqlalchemy.orm import Mapped, mapped_column
+
+from backend.database.base import Base
+
+
+class StorylineVersion(Base):
+    __tablename__ = "storyline_versions"
+    __table_args__ = (
+        {"schema": "memory"},
+    )
+
+    version_id: Mapped[UUID] = mapped_column(
+        PGUUID(as_uuid=True),
+        ForeignKey("memory.memory_versions.id", ondelete="RESTRICT"),
+        primary_key=True,
+    )
+    headline: Mapped[str] = mapped_column(Text, nullable=False)
+    summary: Mapped[str] = mapped_column(Text, nullable=False)
+    status: Mapped[str] = mapped_column(Text, nullable=False)
+    arc_type: Mapped[str | None] = mapped_column(Text, nullable=True)
+    salience: Mapped[int] = mapped_column(SmallInteger, nullable=False)
+    tags: Mapped[list[str]] = mapped_column(ARRAY(Text), nullable=False, default=list)
+    subjects: Mapped[list[dict[str, object]]] = mapped_column(
+        JSONB,
+        nullable=False,
+        default=list,
+        server_default=text("'[]'::jsonb"),
+    )
+    evidence: Mapped[list[dict[str, object]]] = mapped_column(
+        JSONB,
+        nullable=False,
+        default=list,
+        server_default=text("'[]'::jsonb"),
+    )
+    related_storylines: Mapped[list[dict[str, object]]] = mapped_column(
+        JSONB,
+        nullable=False,
+        default=list,
+        server_default=text("'[]'::jsonb"),
+    )
+    callback_condition: Mapped[str | None] = mapped_column(Text, nullable=True)
+    resolution_summary: Mapped[str | None] = mapped_column(Text, nullable=True)
